feat: add components for expiry, state, trading mode, delete unused vol component
