+++ apps/trading/components/market-header/stats/index.ts
@@ -3,3 +3,6 @@ export { Last24hPriceChangeStat } from './last-24h-price-change';
 export { Last24hVolumeChangeStat } from './last-24h-volume-change';
 export { LastTradeStat } from './last-trade';
 export { MarkPriceStat } from './mark-price';
+export { ExpiryStat } from './expiry';
+export { MarketStateStat } from './market-state';
+export { MarketTradingModeStat } from './market-trading-mode';
