--- apps/trading/components/market-state/index.ts
@@ -1,1 +1,0 @@
-export * from './market-state';
