--- apps/trading/components/market-trading-mode/index.ts
@@ -1,1 +1,0 @@
-export * from './market-trading-mode';
