--- apps/trading/components/market-volume/index.ts
@@ -1,1 +1,0 @@
-export * from './market-volume';
